# Optimize performance of Flex.Item