--- a/src/Controls/src/Core/Layout/FlexExtensions.cs
+++ b/src/Controls/src/Core/Layout/FlexExtensions.cs
@@ -7,26 +7,6 @@ namespace Microsoft.Maui.Controls
 {
 	static class FlexExtensions
 	{
-		public static int IndexOf(this Flex.Item parent, Flex.Item child)
-		{
-			var index = -1;
-			foreach (var it in parent)
-			{
-				index++;
-				if (it == child)
-					return index;
-			}
-			return -1;
-		}
-
-		public static void Remove(this Flex.Item parent, Flex.Item child)
-		{
-			var index = parent.IndexOf(child);
-			if (index < 0)
-				return;
-			parent.RemoveAt((uint)index);
-		}
-
 		public static Rect GetFrame(this Flex.Item item)
 		{
 			return new Rect(item.Frame[0], item.Frame[1], item.Frame[2], item.Frame[3]);