--- a/src/Controls/src/Core/Layout/FlexExtensions.cs
+++ b/src/Controls/src/Core/Layout/FlexExtensions.cs
@@ -1,5 +1,4 @@
-﻿#nullable disable
-using Microsoft.Maui.Graphics;
+﻿using Microsoft.Maui.Graphics;
 using Microsoft.Maui.Layouts;
 using Flex = Microsoft.Maui.Layouts.Flex;