--- a/src/Core/src/Layouts/Flex.cs
+++ b/src/Core/src/Layouts/Flex.cs
@@ -228,7 +228,7 @@ namespace Microsoft.Maui.Layouts.Flex
 	/// <summary>
 	/// An item with flexbox properties. Items can also contain other items and be enumerated.
 	/// </summary>
-	class Item : IEnumerable<Item>
+	class Item : List<Item>
 	{
 		/// <summary>
 		/// Gets the frame (x, y, w, h).
@@ -239,7 +239,6 @@ namespace Microsoft.Maui.Layouts.Flex
 		/// <summary>The parent item.</summary>
 		/// <value>The parent item, or null if the item is a root item.</value>
 		public Item? Parent { get; private set; }
-		List<Item>? Children { get; set; }
 		bool ShouldOrderChildren { get; set; }
 
 		///<summary>This property defines how the layout engine will distribute space between and around child items that have been laid out on multiple lines. This property is ignored if the root item does not have its <see cref="P:Microsoft.Maui.Controls.Flex.Item.Wrap" /> property set to Wrap or WrapReverse.</summary>
@@ -392,10 +391,10 @@ namespace Microsoft.Maui.Layouts.Flex
 			Height = height;
 		}
 
-		public void Add(Item child)
+		public new void Add(Item child)
 		{
 			ValidateChild(child);
-			(Children ?? (Children = new List<Item>())).Add(child);
+			base.Add(child);
 			child.Parent = this;
 			ShouldOrderChildren |= child.Order != 0;
 		}
@@ -403,28 +402,17 @@ namespace Microsoft.Maui.Layouts.Flex
 		public void InsertAt(int index, Item child)
 		{
 			ValidateChild(child);
-			(Children ?? (Children = new List<Item>())).Insert(index, child);
+			base.Insert(index, child);
 			child.Parent = this;
 			ShouldOrderChildren |= child.Order != 0;
 		}
 
 		public Item RemoveAt(uint index)
 		{
-			var child = Children![(int)index];
+			var child = this[(int)index];
 			child.Parent = null;
-			Children.RemoveAt((int)index);
+			base.RemoveAt((int)index);
 			return child;
-		}
-
-		public int Count =>
-			(Children?.Count ?? 0);
-
-		public Item ItemAt(int index) =>
-			Children![index];
-
-		public Item this[int index]
-		{
-			get => ItemAt(index);
 		}
 
 		public Item Root
@@ -453,12 +441,6 @@ namespace Microsoft.Maui.Layouts.Flex
 
 		public SelfSizingDelegate? SelfSizing { get; set; }
 
-		IEnumerator IEnumerable.GetEnumerator() =>
-			((IEnumerable<Item>)this).GetEnumerator();
-
-		IEnumerator<Item> IEnumerable<Item>.GetEnumerator() =>
-			(Children ?? global::System.Linq.Enumerable.Empty<Item>()).GetEnumerator();
-
 		void ValidateChild(Item child)
 		{
 			if (this == child)
@@ -469,7 +451,7 @@ namespace Microsoft.Maui.Layouts.Flex
 
 		static void layout_item(Item item, float width, float height)
 		{
-			if (item.Children == null || item.Children.Count == 0)
+			if (item == null || item.Count == 0)
 				return;
 
 			var layout = new flex_layout();
@@ -1021,7 +1003,7 @@ namespace Microsoft.Maui.Layouts.Flex
 						{
 							int prev = indices[j - 1];
 							int curr = indices[j];
-							if (item.Children![prev].Order <= item.Children[curr].Order)
+							if (item[prev].Order <= item[curr].Order)
 							{
 								break;
 							}
@@ -1057,7 +1039,7 @@ namespace Microsoft.Maui.Layouts.Flex
 			}
 
 			public Item child_at(Item item, int i) =>
-				item.Children![(ordered_indices?[i] ?? i)];
+				item[ordered_indices?[i] ?? i];
 
 			public void cleanup()
 			{