Flex.Item was an IEnumerable but internally stored a collection of Children as a List.

It seemed much simpler to just inherit from a list and remove all of the IEnumerable implementation and related extension methods.

Added a benchmark to confirm the suspicion.

Before:

| Method                    | Mean     | Error     | StdDev    | Gen0     | Gen1     | Gen2    | Allocated |
|-------------------------- |---------:|----------:|----------:|---------:|---------:|--------:|----------:|
| LayoutLotsOfItemsWithWrap | 5.038 ms | 0.1739 ms | 0.5074 ms | 171.8750 | 125.0000 | 62.5000 |   1.56 MB |
| LayoutLotsOfItemsNoWrap   | 4.280 ms | 0.0836 ms | 0.1057 ms | 171.8750 | 125.0000 | 70.3125 |   1.56 MB |


After:


| Method                    | Mean     | Error     | StdDev    | Gen0     | Gen1     | Gen2    | Allocated |
|-------------------------- |---------:|----------:|----------:|---------:|---------:|--------:|----------:|
| LayoutLotsOfItemsWithWrap | 4.384 ms | 0.0805 ms | 0.1930 ms | 171.8750 | 156.2500 | 85.9375 |   1.57 MB |
| LayoutLotsOfItemsNoWrap   | 4.331 ms | 0.0845 ms | 0.0829 ms | 171.8750 | 156.2500 | 85.9375 |   1.57 MB |