@rmarinho 2024-03-05
/azp run MAUI-UITests-public